--- src/std_prelude.rs
@@ -1,23 +1,0 @@
-pub use std::convert::TryFrom;
-#[cfg(feature = "ws-client")]
-pub use std::sync::mpsc::Receiver;
-#[cfg(feature = "ws-client")]
-pub use std::sync::mpsc::Sender as ThreadOut;
-
-pub use log::{debug, error, info};
-pub use metadata::RuntimeMetadataPrefixed;
-pub use serde::de::DeserializeOwned;
-pub use serde_json::Value;
-pub use sp_core::crypto::Pair;
-pub use sp_core::storage::StorageKey;
-pub use sp_runtime::traits::{Block, Header};
-pub use sp_runtime::{
-    generic::SignedBlock, traits::IdentifyAccount, AccountId32 as AccountId, MultiSignature,
-    MultiSigner,
-};
-pub use sp_std::prelude::*;
-pub use sp_version::RuntimeVersion;
-
-pub use crate::node_metadata::Metadata;
-pub use crate::rpc::json_req;
-pub use crate::utils::FromHexString;
